requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
Automat==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
constantly==15.1.0
construct==2.10.56
cycler==0.10.0
gast==0.3.3
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.5.2
Markdown==3.2.2
matplotlib==3.3.2
multidict==4.7.6
multitasking==0.0.9
numpy==1.19.2
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.2
Pillow==7.2.0
ply==3.11
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.9.8
PyHamcrest==2.0.2
pyparsing==2.4.7
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.2
seaborn==0.11.0
simplejson==3.17.2
six==1.15.0
tensor==0.3.6
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
Twisted==20.3.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
yahoo-finance==1.4.0
yarl==1.6.0
yfinance==0.1.54
zope.interface==5.1.0
